centro-datos-salud/gateway/requirements.txt
flask
paho-mqtt
grpcio
grpcio-tools
